web_app/requirements.txt
Flask==2.0.1
google-cloud-pubsub==2.8.0
python-dotenv==0.19.1